Move Video component to themes
